fix: resolve PR-29 review comments (clean from main)
