+++ src/content/viajes-en-grupo/en/sicilia-el-alma-del-mediterraneo.mdx
@@ -29,7 +29,7 @@ location: ["Catania, Italy (Europe)", "Palermo, Italy (Europe)", "Syracuse, Ital
   </TimelineItem>
 
   <TimelineItem date="Day 2" title="Catania - Taormina - Catania" image="assets/europa/italia/sicilia/taormina">
-    **Breakfast**. Departure to **Mount Etna** (subject to weather conditions), the highest and most active volcano in Europe (3,345 meters): the bus will arrive at **Sapienza Refuge** at 1,800 meters altitud. Visit to the extinct craters, the so-called **"Crateri Silvestri"**. **Lunch en route**. Continuation to **Taormina**, located on the peak of the picturesque **Mount Tauro** rock (204 m. altitude). Visit to discover the romantic alleys of the city or to visit the famous **Greek Theatre** from where you can enjoy a magnificent panorama of both Etna and the Ionian Sea. Return to the hotel. **Accommodation**.
+    **Breakfast**. Departure to **Mount Etna** (subject to weather conditions), the highest and most active volcano in Europe (3,345 meters): the bus will arrive at **Sapienza Refuge** at 1,800 meters altitude. Visit to the extinct craters, the so-called **"Crateri Silvestri"**. **Lunch en route**. Continuation to **Taormina**, located on the peak of the picturesque **Mount Tauro** rock (204 m. altitude). Visit to discover the romantic alleys of the city or to visit the famous **Greek Theatre** from where you can enjoy a magnificent panorama of both Etna and the Ionian Sea. Return to the hotel. **Accommodation**.
   </TimelineItem>
 
   <TimelineItem date="Day 3" title="Catania - Syracuse - Noto - Ragusa" image="assets/europa/italia/sicilia/ragusa">
@@ -127,7 +127,7 @@ location: ["Catania, Italy (Europe)", "Palermo, Italy (Europe)", "Syracuse, Ital
       - Ancient Greek Theatre
       - Noto Cathedral
       - Archaeological Park of Syracuse, Neapolis and Duomo of Ortigia
-      - Valley of the Templos in Agrigento
+      - Valley of the Temples in Agrigento
       - Villa Romana del Casale
       - Palatine Chapel and Cathedral of Palermo
       - Church of the Martorana in Palermo and Santa Caterina with terrace and Monastery
