+++ src/content/viajes-en-grupo/es/sicilia-el-alma-del-mediterraneo.mdx
@@ -126,7 +126,7 @@ location: ["Catania", "Palermo", "Siracusa", "Agrigento"]
       - Villa Romana del Casale
       - Capilla Palatina y Catedral de Palermo
       - Iglesia de la Martorana en Palermo y de Santa Caterina con terraza y Monasterio
-      - Templo Dorico de Segesta
+      - Templo Dórico de Segesta
       - Claustro benedictino y Catedral de Monreale
       - Catedral Normana (NO TORRES) y Lavadero Medieval en Cefalù
     - Régimen de **media pensión** (bebida no incluida) a partir de la cena del primer día hasta el desayuno del último día.
